+++ framework/i18n/Formatter.php
@@ -48,6 +48,31 @@ use yii\helpers\HtmlPurifier;
 class Formatter extends Component
 {
     /**
+     * @since 2.0.13
+     */
+    const UNIT_SYSTEM_METRIC = 'metric';
+    /**
+     * @since 2.0.13
+     */
+    const UNIT_SYSTEM_IMPERIAL = 'imperial';
+    /**
+     * @since 2.0.13
+     */
+    const FORMAT_WIDTH_LONG = 'long';
+    /**
+     * @since 2.0.13
+     */
+    const FORMAT_WIDTH_SHORT = 'short';
+    /**
+     * @since 2.0.13
+     */
+    const UNIT_LENGTH = 'length';
+    /**
+     * @since 2.0.13
+     */
+    const UNIT_WEIGHT = 'mass';
+
+    /**
      * @var string the text to be displayed when formatting a `null` value.
      * Defaults to `'<span class="not-set">(not set)</span>'`, where `(not set)`
      * will be translated according to [[locale]].
@@ -254,11 +279,89 @@ class Formatter extends Component
      * Defaults to 1024.
      */
     public $sizeFormatBase = 1024;
-
+    /**
+     * @var string default system of measure units. Defaults to [[UNIT_SYSTEM_METRIC]].
+     * Possible values:
+     *  - [[UNIT_SYSTEM_METRIC]]
+     *  - [[UNIT_SYSTEM_IMPERIAL]]
+     * @since 2.0.13
+     */
+    public $systemOfUnits = self::UNIT_SYSTEM_METRIC;
+    /**
+     * @var array configuration of weight and length measurement units.
+     * This array contains the most usable measurement units, but you can change it
+     * in case you have some special requirements.
+     *
+     * For example, you can add smaller measure unit:
+     *
+     * ```php
+     * $this->measureUnits[self::UNIT_LENGTH][self::UNIT_SYSTEM_METRIC] = [
+     *     'nanometer' => 0.000001
+     * ]
+     * ```
+     * @since 2.0.13
+     */
+    public $measureUnits = [
+        self::UNIT_LENGTH => [
+            self::UNIT_SYSTEM_IMPERIAL => [
+                'inch' => 1,
+                'foot' => 12,
+                'yard' => 36,
+                'chain' => 792,
+                'furlong' => 7920,
+                'mile' => 63360,
+            ],
+            self::UNIT_SYSTEM_METRIC => [
+                'millimeter' => 1,
+                'centimeter' => 10,
+                'meter' => 1000,
+                'kilometer' => 1000000,
+            ],
+        ],
+        self::UNIT_WEIGHT => [
+            self::UNIT_SYSTEM_IMPERIAL => [
+                'grain' => 1,
+                'drachm' => 27.34375,
+                'ounce' => 437.5,
+                'pound' => 7000,
+                'stone' => 98000,
+                'quarter' => 196000,
+                'hundredweight' => 784000,
+                'ton' => 15680000,
+            ],
+            self::UNIT_SYSTEM_METRIC => [
+                'gram' => 1,
+                'kilogram' => 1000,
+                'ton' => 1000000,
+            ],
+        ],
+    ];
+    /**
+     * @var array The base units that are used as multipliers for smallest possible unit from [[measureUnits]]
+     * @since 2.0.13
+     */
+    public $baseUnits = [
+        self::UNIT_LENGTH => [
+            self::UNIT_SYSTEM_IMPERIAL => 12, // 1 feet = 12 inches
+            self::UNIT_SYSTEM_METRIC => 1000, // 1 meter = 1000 millimeters
+        ],
+        self::UNIT_WEIGHT => [
+            self::UNIT_SYSTEM_IMPERIAL => 7000, // 1 pound = 7000 grains
+            self::UNIT_SYSTEM_METRIC => 1000, // 1 kilogram = 1000 grams
+        ],
+    ];
     /**
      * @var bool whether the [PHP intl extension](http://php.net/manual/en/book.intl.php) is loaded.
      */
     private $_intlLoaded = false;
+    /**
+     * @var \ResourceBundle cached ResourceBundle object used to read unit translations
+     */
+    private $_resourceBundle;
+    /**
+     * @var array cached unit translation patterns
+     */
+    private $_unitMessages = [];
 
 
     /**
@@ -1205,7 +1308,7 @@ class Formatter extends Component
             return $this->nullDisplay;
         }
 
-        list($params, $position) = $this->formatSizeNumber($value, $decimals, $options, $textOptions);
+        list($params, $position) = $this->formatNumber($value, $decimals, 4, $this->sizeFormatBase, $options, $textOptions);
 
         if ($this->sizeFormatBase == 1024) {
             switch ($position) {
@@ -1261,7 +1364,7 @@ class Formatter extends Component
             return $this->nullDisplay;
         }
 
-        list($params, $position) = $this->formatSizeNumber($value, $decimals, $options, $textOptions);
+        list($params, $position) = $this->formatNumber($value, $decimals, 4, $this->sizeFormatBase, $options, $textOptions);
 
         if ($this->sizeFormatBase == 1024) {
             switch ($position) {
@@ -1296,31 +1399,209 @@ class Formatter extends Component
         }
     }
 
+    /**
+     * Formats the value as a length in human readable form for example `12 meters`.
+     *
+     * @param double|int $value value to be formatted.
+     * @param double $baseUnit unit of value as the multiplier of the smallest unit.
+     * @param string $unitSystem either [[UNIT_SYSTEM_METRIC]] or [[UNIT_SYSTEM_IMPERIAL]]. When `null`, property [[systemOfUnits]] will be used.
+     * @param int $decimals the number of digits after the decimal point.
+     * @param array $options optional configuration for the number formatter. This parameter will be merged with [[numberFormatterOptions]].
+     * @param array $textOptions optional configuration for the number formatter. This parameter will be merged with [[numberFormatterTextOptions]].
+     * @return string the formatted result.
+     * @throws InvalidParamException if the input value is not numeric or the formatting failed.
+     * @throws InvalidConfigException when INTL is not installed or does not contain required information.
+     * @see asLength
+     * @since 2.0.13
+     * @author John Was <janek.jan@gmail.com>
+     */
+    public function asLength($value, $baseUnit = null, $unitSystem = null, $decimals = null, $options = [], $textOptions = [])
+    {
+        return $this->formatUnit(self::UNIT_LENGTH, self::FORMAT_WIDTH_LONG, $value, $baseUnit, $unitSystem, $decimals, $options, $textOptions);
+    }
+
+    /**
+     * Formats the value as a length in human readable form for example `12 m`.
+     *
+     * This is the short form of [[asLength]].
+     *
+     * @param double|int $value value to be formatted.
+     * @param double $baseUnit unit of value as the multiplier of the smallest unit
+     * @param string $unitSystem either [[UNIT_SYSTEM_METRIC]] or [[UNIT_SYSTEM_IMPERIAL]]. When `null`, property [[systemOfUnits]] will be used.
+     * @param int $decimals the number of digits after the decimal point.
+     * @param array $options optional configuration for the number formatter. This parameter will be merged with [[numberFormatterOptions]].
+     * @param array $textOptions optional configuration for the number formatter. This parameter will be merged with [[numberFormatterTextOptions]].
+     * @return string the formatted result.
+     * @throws InvalidParamException if the input value is not numeric or the formatting failed.
+     * @throws InvalidConfigException when INTL is not installed or does not contain required information.
+     * @see asLength
+     * @since 2.0.13
+     * @author John Was <janek.jan@gmail.com>
+     */
+    public function asShortLength($value, $baseUnit = null, $unitSystem = null, $decimals = null, $options = [], $textOptions = [])
+    {
+        return $this->formatUnit(self::UNIT_LENGTH, self::FORMAT_WIDTH_SHORT, $value, $baseUnit, $unitSystem, $decimals, $options, $textOptions);
+    }
+
+    /**
+     * Formats the value as a weight in human readable form for example `12 kilograms`.
+     *
+     * @param double|int $value value to be formatted.
+     * @param double $baseUnit unit of value as the multiplier of the smallest unit
+     * @param string $unitSystem either [[UNIT_SYSTEM_METRIC]] or [[UNIT_SYSTEM_IMPERIAL]]. When `null`, property [[systemOfUnits]] will be used.
+     * @param int $decimals the number of digits after the decimal point.
+     * @param array $options optional configuration for the number formatter. This parameter will be merged with [[numberFormatterOptions]].
+     * @param array $textOptions optional configuration for the number formatter. This parameter will be merged with [[numberFormatterTextOptions]].
+     * @return string the formatted result.
+     * @throws InvalidParamException if the input value is not numeric or the formatting failed.
+     * @throws InvalidConfigException when INTL is not installed or does not contain required information.
+     * @since 2.0.13
+     * @author John Was <janek.jan@gmail.com>
+     */
+    public function asWeight($value, $baseUnit = null, $unitSystem = null, $decimals = null, $options = [], $textOptions = [])
+    {
+        return $this->formatUnit(self::UNIT_WEIGHT, self::FORMAT_WIDTH_LONG, $value, $baseUnit, $unitSystem, $decimals, $options, $textOptions);
+    }
+
+    /**
+     * Formats the value as a weight in human readable form for example `12 kg`.
+     *
+     * This is the short form of [[asWeight]].
+     *
+     * @param double|int $value value to be formatted.
+     * @param double $baseUnit unit of value as the multiplier of the smallest unit
+     * @param string $unitSystem either [[UNIT_SYSTEM_METRIC]] or [[UNIT_SYSTEM_IMPERIAL]]. When `null`, property [[systemOfUnits]] will be used.
+     * @param int $decimals the number of digits after the decimal point.
+     * @param array $options optional configuration for the number formatter. This parameter will be merged with [[numberFormatterOptions]].
+     * @param array $textOptions optional configuration for the number formatter. This parameter will be merged with [[numberFormatterTextOptions]].
+     * @return string the formatted result.
+     * @throws InvalidParamException if the input value is not numeric or the formatting failed.
+     * @throws InvalidConfigException when INTL is not installed or does not contain required information.
+     * @since 2.0.13
+     * @author John Was <janek.jan@gmail.com>
+     */
+    public function asShortWeight($value, $baseUnit = null, $unitSystem = null, $decimals = null, $options = [], $textOptions = [])
+    {
+        return $this->formatUnit(self::UNIT_WEIGHT, self::FORMAT_WIDTH_SHORT, $value, $baseUnit, $unitSystem, $decimals, $options, $textOptions);
+    }
+
+    /**
+     * @param string $unitType one of [[UNIT_WEIGHT]], [[UNIT_LENGTH]]
+     * @param string $unitFormat one of [[FORMAT_WIDTH_SHORT]], [[FORMAT_WIDTH_LONG]]
+     * @param double|int $value to be formatted
+     * @param double $baseUnit unit of value as the multiplier of the smallest unit
+     * @param string $unitSystem either [[UNIT_SYSTEM_METRIC]] or [[UNIT_SYSTEM_IMPERIAL]]. When `null`, property [[systemOfUnits]] will be used.
+     * @param int $decimals the number of digits after the decimal point.
+     * @param array $options optional configuration for the number formatter. This parameter will be merged with [[numberFormatterOptions]].
+     * @param array $textOptions optional configuration for the number formatter. This parameter will be merged with [[numberFormatterTextOptions]].
+     * @return string
+     * @throws InvalidConfigException when INTL is not installed or does not contain required information
+     */
+    private function formatUnit($unitType, $unitFormat, $value, $baseUnit, $unitSystem, $decimals, $options, $textOptions)
+    {
+        if ($value === null) {
+            return $this->nullDisplay;
+        }
+        if ($unitSystem === null) {
+            $unitSystem = $this->systemOfUnits;
+        }
+        if ($baseUnit === null) {
+            $baseUnit = $this->baseUnits[$unitType][$unitSystem];
+        }
+
+        $multipliers = array_values($this->measureUnits[$unitType][$unitSystem]);
+
+        list($params, $position) = $this->formatNumber($value * $baseUnit, $decimals, null, $multipliers, $options, $textOptions);
+
+        $message = $this->getUnitMessage($unitType, $unitFormat, $unitSystem, $position);
+
+        return (new \MessageFormatter($this->locale, $message))->format([
+            '0' => $params['nFormatted'],
+            'n' => $params['n'],
+        ]);
+    }
+
+    /**
+     * @param string $unitType one of [[UNIT_WEIGHT]], [[UNIT_LENGTH]]
+     * @param string $unitFormat one of [[FORMAT_WIDTH_SHORT]], [[FORMAT_WIDTH_LONG]]
+     * @param string $system either [[UNIT_SYSTEM_METRIC]] or [[UNIT_SYSTEM_IMPERIAL]]. When `null`, property [[systemOfUnits]] will be used.
+     * @param int $position internal position of size unit
+     * @return string
+     * @throws InvalidConfigException when INTL is not installed or does not contain required information
+     */
+    private function getUnitMessage($unitType, $unitFormat, $system, $position)
+    {
+        if (isset($this->_unitMessages[$unitType][$system][$position])) {
+            return $this->_unitMessages[$unitType][$system][$position];
+        }
+        if (!$this->_intlLoaded) {
+            throw new InvalidConfigException('Format of ' . $unitType . ' is only supported when PHP intl extension is installed.');
+        }
+
+        if ($this->_resourceBundle === null) {
+            $this->_resourceBundle = new \ResourceBundle($this->locale, 'ICUDATA-unit');
+        }
+        $unitNames = array_keys($this->measureUnits[$unitType][$system]);
+        $bundleKey = 'units' . ($unitFormat === self::FORMAT_WIDTH_SHORT ? 'Short' : '');
+
+        $unitBundle = $this->_resourceBundle[$bundleKey][$unitType][$unitNames[$position]];
+        if ($unitBundle === null) {
+            throw new InvalidConfigException('Current version of ICU data does not contain information about unit type "' . $unitType . '" and unit measure "' . $unitNames[$position] . '. Check system requirements.');
+        }
+
+        $message = [];
+        foreach ($unitBundle as $key => $value) {
+            if ($key === 'dnam') {
+                continue;
+            }
+            $message[] = "$key{{$value}}";
+        }
+        return $this->_unitMessages[$unitType][$system][$position] = '{n, plural, '.implode(' ', $message).'}';
+    }
 
     /**
      * Given the value in bytes formats number part of the human readable form.
      *
      * @param string|int|float $value value in bytes to be formatted.
      * @param int $decimals the number of digits after the decimal point
+     * @param int $maxPosition maximum internal position of size unit, ignored if $formatBase is an array
+     * @param array|int $formatBase the base at which each next unit is calculated, either 1000 or 1024, or an array
      * @param array $options optional configuration for the number formatter. This parameter will be merged with [[numberFormatterOptions]].
      * @param array $textOptions optional configuration for the number formatter. This parameter will be merged with [[numberFormatterTextOptions]].
      * @return array [parameters for Yii::t containing formatted number, internal position of size unit]
      * @throws InvalidParamException if the input value is not numeric or the formatting failed.
      */
-    private function formatSizeNumber($value, $decimals, $options, $textOptions)
+    private function formatNumber($value, $decimals, $maxPosition, $formatBase, $options, $textOptions)
     {
         $value = $this->normalizeNumericValue($value);
 
         $position = 0;
+        if (is_array($formatBase)) {
+            $maxPosition = count($formatBase) - 1;
+        }
         do {
-            if (abs($value) < $this->sizeFormatBase) {
-                break;
-            }
-            $value /= $this->sizeFormatBase;
-            $position++;
-        } while ($position < 5);
+            if (is_array($formatBase)) {
+                if (!isset($formatBase[$position + 1])) {
+                    break;
+                }
 
-        // no decimals for bytes
+                if (abs($value) < $formatBase[$position + 1]) {
+                    break;
+                }
+            } else {
+                if (abs($value) < $formatBase) {
+                    break;
+                }
+                $value = $value / $formatBase;
+            }
+            $position++;
+        } while ($position < $maxPosition + 1);
+
+        if (is_array($formatBase) && $position !== 0) {
+            $value /= $formatBase[$position];
+        }
+
+        // no decimals for smallest unit
         if ($position === 0) {
             $decimals = 0;
         } elseif ($decimals !== null) {